+++ src/action/getStyles.tsx
@@ -1,26 +1,6 @@
 import React from 'react'
 import type { ExecProps } from './index'
-import classNames from 'classnames'
-
-import {
-  resultPopupButton,
-  resultTable,
-  resultTableContainer,
-  resultTableHeader,
-  resultTableBody,
-  resultTableProperty,
-  resultTableValue,
-  resultTableCopy,
-} from '@/components/menu/ResultPopup.module.css'
-import { Icon } from '@/components/Icon'
-
-const toName = (str: string) => {
-  return str.replace(/([A-Z])/g, ' $1').replace(/^./, (s) => s.toUpperCase())
-}
-
-const toProp = (str: string) => {
-  return str.replace(/([A-Z])/g, '-$1').replace(/^./, (s) => s.toLowerCase())
-}
+import { TextStyle } from '@/components/result/TextStyle'
 
 export const GetStyles = {
   async execute({ target }: ExecProps) {
@@ -29,59 +9,10 @@ export const GetStyles = {
     }
 
     const styles = getFontCSS(target as HTMLElement)
-    const ret = Object.entries(styles).map(([key, value]) => ({
-      key,
-      value,
-    }))
 
-    const cssCopy = () => {
-      const copyText = ret
-        .map((item) => `${toProp(item.key)}: ${item.value};`)
-        .join('\n')
-      navigator.clipboard.writeText(copyText)
-    }
-
-    return (
-      <div className={resultTableContainer}>
-        <button
-          className={classNames(resultTableCopy, resultPopupButton)}
-          onClick={cssCopy}
-        >
-          <Icon name="copy" />
-        </button>
-        <table className={resultTable}>
-          <thead className={resultTableHeader}>
-            <tr key="table-header">
-              <th>Property</th>
-              <th>Value</th>
-            </tr>
-          </thead>
-          <tbody className={resultTableBody}>
-            {ret.map((item) => {
-              return (
-                <tr key={item.key}>
-                  <td className={resultTableProperty}>{toName(item.key)}</td>
-                  <td className={resultTableValue}>{item.value}</td>
-                </tr>
-              )
-            })}
-          </tbody>
-        </table>
-      </div>
-    )
+    return <TextStyle styles={styles} />
   },
 }
-
-type FontCSS = Pick<
-  CSSStyleDeclaration,
-  | 'fontFamily'
-  | 'fontWeight'
-  | 'fontStyle'
-  | 'fontSize'
-  | 'color'
-  | 'lineHeight'
-  | 'letterSpacing'
->
 
 const CANVAS_OPTIONS = {
   fillStyle: 'rgb(0,0,0)',
