+++ src/typings.d.ts
@@ -4,3 +4,14 @@ declare module '*.svg' {
   const content: any
   export default content
 }
+
+declare type FontCSS = Pick<
+  CSSStyleDeclaration,
+  | 'fontFamily'
+  | 'fontWeight'
+  | 'fontStyle'
+  | 'fontSize'
+  | 'color'
+  | 'lineHeight'
+  | 'letterSpacing'
+>
